% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_srtm.R
\name{download_srtm}
\alias{download_srtm}
\title{\code{download_srtm}
Download the SRTM maps and unzip them into the specified folder. Helper function to get_elevation_srtm()}
\usage{
download_srtm(url_file, destfile, exdir_srtm_hgt = "SRTM")
}
\arguments{
\item{url_file}{Url where the SRTM maps are stored}

\item{destfile}{Given name of the file of the downloaded map}

\item{exdir_srtm_hgt}{Default: "SRTM/". Folder where the maps are downloaded and unzipped. The zip file is deleted and only the .hgt files are kept}
}
\value{
Nothing to return.
}
\description{
\code{download_srtm}
Download the SRTM maps and unzip them into the specified folder. Helper function to get_elevation_srtm()
}
\examples{
library(SRTM)
download_srtm(url_file = SRTM_files_info$full_path[1],
              destfile = SRTM_files_info$name_file[1] )
# See downloaded file in SRTM/N00E006.hgt

}
\seealso{
\code{\link{SRTM}}
}
\author{
Javier Saez Gallego
}
