% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_srtm_raster.R
\name{get_srtm_raster}
\alias{get_srtm_raster}
\title{\code{get_srtm_raster}
Given a pair of coordinates, find which SRTM file to download. Basically, find all the surrounding files and return them as a merged raster.}
\usage{
get_srtm_raster(lon, lat, n = 1, exdir_srtm_hgt = "SRTM")
}
\arguments{
\item{lon}{Longitude given in WGS84}

\item{lat}{Latitude given in WGS84}

\item{n}{Numbers of map tiles to gather around the specified coordinates. If n=1 (default), 9 tiles are merged, if n=2, 25 tiles are merged. In general, (2*n+1)^2 tiles are downloaded and merged so the computational time increases exponentially too. The length of the side of each tile is roughly 1 degree (approx 111km)}

\item{exdir_srtm_hgt}{Default: "SRTM/". Folder where the maps are downloaded and unzipped. The zip file is deleted and only the .hgt files are kept. The folder is created if it does not exist.}
}
\value{
Raster image containing merged SRTM files.
}
\description{
\code{get_srtm_raster}
Given a pair of coordinates, find which SRTM file to download. Basically, find all the surrounding files and return them as a merged raster.
}
\examples{
# Download and plot maps around Mout Everest
library(SRTM)
everest =  get_srtm_raster(lon = 86.922623, lat = 27.986065 , n=0, exdir_srtm_hgt = "SRTM")
# Takes a while the first time
# See downloaded .hgt files in SRTM
raster::plot(everest)

}
\seealso{
\code{\link{SRTM}}
}
\author{
Javier Saez Gallego
}
