% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRTM.R
\docType{package}
\name{SRTM}
\alias{SRTM}
\alias{SRTM-package}
\title{SRTM}
\description{
Collection of functions that facilitate the use or SRTM maps in R. Download
maps around a specified area and returns a merged raster.
}
\details{
Information about the SRTM maps is available \href{https://wiki.openstreetmap.org/wiki/SRTM}{here} and \href{https://www2.jpl.nasa.gov/srtm/}{here}.
See a map of the available areas: \url{https://dds.cr.usgs.gov/srtm/version2_1/Documentation/Continent_def.gif}

The folowing functions are included:
\itemize{
\item \code{\link{get_srtm_raster}}
\item \code{\link{download_srtm}}
}

Also, there is a dataset \code{\link{SRTM_files_info}}
}
\author{
Javier Saez Gallego <javiersaezgallego at gmail dot com>
}
